[Core] Split LoRA layers (#24574)
